AUTO-004: Relocated Test Files and Custom Commands
