+++ cypress/e2e/APITests/apiCommands.d.ts
@@ -1,0 +1,10 @@
+declare namespace Cypress{
+    interface chainable{
+        
+        /**
+         * @description This cmd deletes the UI-created article from API
+         * @example cy.deleteCreatedArticle()
+         */
+                deleteCreatedArticle():Chainable<Element>
+        
+    }}
